refactorization : deplacement de UseAnimation dans le dossier cistom-hooks
